Add missing autoconf dependency
Line 27 uses autoconf, but the package isn't installed in the lines above. On a fresh install of Ubuntu 20.04, the autoconf package isn't installed, so line 27 gave an error "Command 'autoconf' not found, but can be installed with: sudo apt install autoconf". Simply including the autoconf package in the install lines solves the problem (and will be ignored if autoconf is already installed).